refactor(#zimic): split fetch and url utilities
