+++ packages/zimic/src/interceptor/http/interceptor/HttpInterceptorClient.ts
@@ -7,7 +7,7 @@ import {
   HttpServiceSchemaPath,
 } from '@/http/types/schema';
 import { Default, PossiblePromise } from '@/types/utils';
-import { ExtendedURL, joinURL } from '@/utils/fetch';
+import { joinURL, ExtendedURL } from '@/utils/urls';
 
 import HttpInterceptorWorker from '../interceptorWorker/HttpInterceptorWorker';
 import LocalHttpInterceptorWorker from '../interceptorWorker/LocalHttpInterceptorWorker';
