+++ packages/zimic/src/interceptor/http/interceptor/__tests__/shared/baseURLs.ts
@@ -1,7 +1,7 @@
 import { beforeEach, expect, expectTypeOf, it, vi } from 'vitest';
 
 import { promiseIfRemote } from '@/interceptor/http/interceptorWorker/__tests__/utils/promises';
-import { ExtendedURL, joinURL } from '@/utils/fetch';
+import { joinURL, ExtendedURL } from '@/utils/urls';
 import { usingHttpInterceptor } from '@tests/utils/interceptors';
 
 import { SUPPORTED_BASE_URL_PROTOCOLS } from '../../HttpInterceptorClient';
