+++ packages/zimic/src/interceptor/http/requestHandler/__tests__/shared/default.ts
@@ -8,8 +8,8 @@ import { HttpInterceptorType } from '@/interceptor/http/interceptor/types/option
 import { promiseIfRemote } from '@/interceptor/http/interceptorWorker/__tests__/utils/promises';
 import HttpInterceptorWorker from '@/interceptor/http/interceptorWorker/HttpInterceptorWorker';
 import LocalHttpInterceptorWorker from '@/interceptor/http/interceptorWorker/LocalHttpInterceptorWorker';
-import { joinURL } from '@/utils/fetch';
 import { waitForDelay } from '@/utils/time';
+import { joinURL } from '@/utils/urls';
 import { createInternalHttpInterceptor } from '@tests/utils/interceptors';
 
 import NoResponseDefinitionError from '../../errors/NoResponseDefinitionError';
