+++ packages/zimic/src/interceptor/http/requestHandler/__tests__/shared/types.ts
@@ -1,7 +1,7 @@
 import { HttpSchema } from '@/http/types/schema';
 import { HttpInterceptorPlatform, HttpInterceptorType } from '@/interceptor/http/interceptor/types/options';
 import { PossiblePromise } from '@/types/utils';
-import { ExtendedURL } from '@/utils/fetch';
+import { ExtendedURL } from '@/utils/urls';
 
 export interface SharedHttpRequestHandlerTestOptions {
   platform: HttpInterceptorPlatform;
